introduce project references to ensure that shared package(s) are built whenever we build the server.
modify build command for server to use typescript's build mode. This works with project references to ensure that "referenced" projects are built whenever the dependent project is built.

In our case the packages/shared will be build whenever we build the server

introduce conditional exports to only point to our built cjs when used with node